Use browserify-hmr instead of webpack for dev
Browserify-hmr allows us to update the websocket url at runtime which is
great when accessing the app on a different hostname. We don't need
webpack dev server anymore or the processhtml task since the js file is
actually written onto disk instead of being in-memory of the
webpack dev server.